+++ assets/release_notes.md
@@ -6,6 +6,7 @@
 - Fix broken documentation link
 - Reduce frequency of notification checks
 - Updated translations
+- Add image cropping functionality
 
 ### 0.18.1 - April 2025
 ---